requirements.txt
# ================================================
# BuryatVPN Dependencies
# ================================================

# Core Framework
aiogram>=3.1.1
Flask>=3.0.0
SQLAlchemy>=2.0.0
alembic>=1.12.0

# Async & Database
aiosqlite>=0.19.0
asyncpg>=0.28.0
redis>=5.0.0
aioredis>=2.0.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP & API
aiohttp>=3.9.1
requests>=2.31.0
httpx>=0.25.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Payment Systems
yookassa>=3.0.0
aiocryptopay>=1.3.0

# X-UI Integration
py3xui>=1.0.0

# Utilities
loguru>=0.7.2
qrcode[pil]>=7.4.2
Pillow>=10.0.0
python-dateutil>=2.8.2

# Monitoring & Metrics
prometheus-client>=0.18.0
psutil>=5.9.0

# Web Framework Extensions
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0
Flask-Talisman>=1.1.0
Flask-JWT-Extended>=4.6.0

# CLI & Development
click>=8.1.0
python-multipart>=0.0.6

# Validation & Serialization
marshmallow>=3.20.0
email-validator>=2.0.0
